deleted: unwanted routes in router.jsx and moved axiosDefaults.jsx to API dir
